Build successfully on Mac arm64. Bump version to 0.2.
`brew install llvm`, then `make static`.

Automatically use brew's clang on macOS ARM; system clang lacks the -fopenmp flag. Couldn't get GCC to build it successfully on ARM dur to failure to find gomp during linking. Gomp appears not to be included with brew's GCC, nor is it packaged separately.

Rename GCC var to COMPILER now that we default to clang in some cases. (Can't use "CC" because that's a make built-in.)